fix init docker script
the init script of failed a restart if an image was running.
Problem was, that restart just calls stop and then start.
But if docker as to work before shutdown, it is still running when
start is called.
The workaround is, to ship our own init script that waits for docker to
stop before starting.